Implement pooled mode reusing the same UDP Socket (Very Experimental)
- Abstract Socket out to ISocket to support both normal queries and pooled queries
- Implement pooling logic
- Implement UDPDemultiplexer, a helper to redirect incoming packets on the shared socket to the appropriate awaiting query
- Update tests/examples with pooling

To Do:
Looking into the ByteReader code, it seems there is a lot of little optimizations that could be made. ByteReader is recreated more then once, and there are a few points where byte arrays are spliced together. It might make sense to use Memory<byte> all the way down instead, and a shared bytereader for each query.